Fix bad bug that allows .svn files to be in a repo
This fix just checks to see if .svn files are present in either the
master or local-hedgehog branch and refuses to merge the two if
they are present.
But this step is only run during the initial merge, and so this problem could
still crop up if .svn files are committed to git (or svn if svn allows that)
at some later point. So probably need to fix that too.